+++ packages/docs-snippets/src/snippets/designation.ts
@@ -15,7 +15,7 @@ import * as ledger from '@midnight-ntwrk/ledger-v8';
 import { Buffer } from 'buffer';
 import * as rx from 'rxjs';
 import { initWalletWithSeed } from '../utils.ts';
-import { generateRandomSeed } from '@midnight-ntwrk/wallet-sdk-hd';
+import { generateRandomSeed } from '@midnight-ntwrk/wallet-sdk';
 
 const sender = await initWalletWithSeed(
   Buffer.from('0000000000000000000000000000000000000000000000000000000000000001', 'hex'),