+++ packages/dust-wallet/src/v1/CoinsAndBalances.ts
@@ -13,9 +13,15 @@
 import * as ledger from '@midnight-ntwrk/ledger-v8';
 import { DateOps } from '@midnight-ntwrk/wallet-sdk-utilities';
 import { pipe, Array as Arr, Order } from 'effect';
-import { CoreWallet } from './CoreWallet.js';
-import { KeysCapability } from './Keys.js';
-import { DustGenerationDetails, DustGenerationInfo, Dust, DustFullInfo, UtxoWithMeta } from './types/Dust.js';
+import { type CoreWallet } from './CoreWallet.js';
+import { type KeysCapability } from './Keys.js';
+import {
+  type DustGenerationDetails,
+  type DustGenerationInfo,
+  type Dust,
+  type DustFullInfo,
+  type UtxoWithMeta,
+} from './types/Dust.js';
 
 export type Balance = bigint;
 
@@ -25,8 +31,8 @@ export type CoinWithValue<TToken> = {
 };
 
 /**
- * Type describing a Night UTxO together with details of estimated Dust generation.
- * It is meant to be primarily used for fee estimation of Dust registration transaction
+ * Type describing a Night UTxO together with details of estimated Dust generation. It is meant to be primarily used for
+ * fee estimation of Dust registration transaction
  */
 export type UtxoWithFullDustDetails = Readonly<{
   utxo: UtxoWithMeta;
@@ -46,17 +52,17 @@ export type CoinsAndBalancesCapability<TState> = {
   getAvailableCoinsWithGeneratedDust(state: TState, currentTime: Date): ReadonlyArray<CoinWithValue<Dust>>;
   getGenerationInfo(state: TState, coin: Dust): DustGenerationInfo | undefined;
 
-  /**
-   * Splits provided Night utxos into the ones that will be used as inputs in the guaranteed and fallible sections
-   */
+  /** Splits provided Night utxos into the ones that will be used as inputs in the guaranteed and fallible sections */
   splitNightUtxos(nightUtxos: ReadonlyArray<UtxoWithFullDustDetails>): {
     guaranteed: ReadonlyArray<UtxoWithFullDustDetails>;
     fallible: ReadonlyArray<UtxoWithFullDustDetails>;
   };
 
   /**
-   * Estimate how much Dust would be available to use if the Utxos provided were used for Dust generation from their beginning.
-   * This function is particularly useful for the purpose of registering for Dust generation and selecting the Utxo to be used for paying fees and approving the registration itself.
+   * Estimate how much Dust would be available to use if the Utxos provided were used for Dust generation from their
+   * beginning. This function is particularly useful for the purpose of registering for Dust generation and selecting
+   * the Utxo to be used for paying fees and approving the registration itself.
+   *
    * @param state Current state of the wallet
    * @param nightUtxos Existing Night utxos
    * @param currentTime Current time
@@ -155,9 +161,7 @@ export const makeDefaultCoinsAndBalancesCapability = (
     );
   };
 
-  /**
-   * Create a fake generation info for a given Utxo. It allows to estimate the Dust generation from it
-   */
+  /** Create a fake generation info for a given Utxo. It allows to estimate the Dust generation from it */
   const fakeGenerationInfo = (utxo: UtxoWithMeta, dustPublicKey: ledger.DustPublicKey): DustGenerationInfo => {
     return {
       value: utxo.value,
@@ -167,9 +171,7 @@ export const makeDefaultCoinsAndBalancesCapability = (
     };
   };
 
-  /**
-   * Create a fake dust coin for a given Utxo. It allows to estimate full details of the Dust generation from it
-   */
+  /** Create a fake dust coin for a given Utxo. It allows to estimate full details of the Dust generation from it */
   const fakeDustToken = (dustPublicKey: ledger.DustPublicKey, utxo: UtxoWithMeta): Dust => ({
     initialValue: 0n,
     owner: dustPublicKey,