+++ packages/indexer-client/src/effect/QueryClient.ts
@@ -10,8 +10,8 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-import { Effect, Context } from 'effect';
-import { ClientError, ServerError } from '@midnight-ntwrk/wallet-sdk-utilities/networking';
+import { type Effect, Context } from 'effect';
+import { type ClientError, type ServerError } from '@midnight-ntwrk/wallet-sdk-utilities/networking';
 import type { Query } from './Query.js';
 
 export class QueryClient extends Context.Tag('@midnight-ntwrk/indexer-client#QueryClient')<