+++ packages/node-client/src/gen/augment-api-query.ts
@@ -11,10 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 /* eslint-disable */
-/**
- * Auto-generated with scripts/generate-types.ts
- * Then manually modified to only expose types and endpoints of interest
- */
+/** Auto-generated with scripts/generate-types.ts Then manually modified to only expose types and endpoints of interest */
 
 // import type lookup before we augment - in some environments
 // this is required to allow for ambient/previous definitions