+++ packages/runtime/src/abstractions/test/variant.test.ts
@@ -11,18 +11,24 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 import { ProtocolVersion } from '@midnight-ntwrk/wallet-sdk-abstractions';
-import { HList, Poly } from '@midnight-ntwrk/wallet-sdk-utilities';
+import { type HList, type Poly } from '@midnight-ntwrk/wallet-sdk-utilities';
 import type { Expect, Equal, CanAssign } from '@midnight-ntwrk/wallet-sdk-utilities/types';
 import { describe, expect, it } from 'vitest';
 import {
-  InterceptingRunningVariant,
+  type InterceptingRunningVariant,
   InterceptingVariant,
   Numeric,
   NumericMultiplier,
   NumericRange,
   NumericRangeMultiplier,
 } from '../../testing/variants.js';
-import { makeVersionedRecord, RunningVariant, RunningVariantOf, StateOf, VersionedVariant } from '../Variant.js';
+import {
+  makeVersionedRecord,
+  type RunningVariant,
+  type RunningVariantOf,
+  type StateOf,
+  type VersionedVariant,
+} from '../Variant.js';
 
 describe('Variant', () => {
   it('infers its state correctly', () => {