+++ packages/shielded-wallet/src/v1/V1Builder.ts
@@ -42,7 +42,7 @@ import { type KeysCapability, makeDefaultKeysCapability } from './Keys.js';
 import { type CoinSelection, chooseCoin } from '@midnight-ntwrk/wallet-sdk-capabilities';
 import { CoreWallet, PublicKeys } from './CoreWallet.js';
 import {
-  DefaultTransactionHistoryConfiguration,
+  type DefaultTransactionHistoryConfiguration,
   makeDefaultTransactionHistoryService,
   type TransactionHistoryService,
 } from './TransactionHistory.js';
@@ -527,9 +527,7 @@ declare namespace V1Builder {
     readonly keysCapability: (configuration: TConfig, getContext: () => TContext) => KeysCapability<CoreWallet>;
   };
 
-  /**
-   * The internal build state of {@link V1Builder}.
-   */
+  /** The internal build state of {@link V1Builder}. */
   type FullBuildState<TConfig, TContext, TSerialized, TSyncUpdate, TTransaction, TStartAux> = Types.Simplify<
     HasSync<TConfig, TContext, TSyncUpdate, TStartAux> &
       HasSerialization<TConfig, TContext, TSerialized> &
@@ -552,9 +550,7 @@ declare namespace V1Builder {
       | undefined;
   };
 
-  /**
-   * Utility interface that manages the type variance of {@link V1Builder}.
-   */
+  /** Utility interface that manages the type variance of {@link V1Builder}. */
   interface Variance<R> {
     readonly [V1BuilderSymbol.typeId]: {
       readonly _R: Types.Covariant<R>;
@@ -575,9 +571,7 @@ const isBuildStateFull = <TConfig, TContext, TSerialized, TSyncUpdate, TTransact
     'keysCapability',
     'transactionHistoryService',
   ] as const;
-  /**
-   * This type will fail compilation if any key is omitted, letting the `isFull` check work properly
-   */
+  /** This type will fail compilation if any key is omitted, letting the `isFull` check work properly */
   type _1 = Expect<
     Equal<keyof V1Builder.FullBuildState<never, never, never, never, never, never>, ItemType<typeof allBuildStateKeys>>
   >;